reorganized WeightedMDS and weight_to_matrices
